Activate critical state when battery is less than or eqal
At the current moment, the deck will be suspended when the battery is less than the threshold. Instead of checking whether the battery is less than the target, we should check whether it's less than or equal to.

I believe this would make more sense, as the slider in the settings for AutoSuspend displays that the critical threshold might be 5%, leading the user to expect their deck to be suspended at 5% battery. Instead, it will suspend at 4%, which might lead some users into thinking that the plugin didn't work.